+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/gui/items/CycleSortItem.kt
@@ -15,6 +15,7 @@ class CycleSortItem(wpGUI: WaypointsGUI, onCycle: suspend (OverviewSort) -> Unit
             launch { wpGUI.viewerData.setSortBy(it) }
             onCycle(it)
           }
+          wpGUI.playSound { clickNormal }
         }) {
 
   init {